Wifi connect (#5)
* cleanup ui message. log to stdout

* css dir needs to be called static (flask)

* ui updates

set correct project link at the bottom of the form.
fix uid/piid collection

* cleanup

fix syntax error when opening a file.
make the message text smaller

* init is not necessary mount shared secret instead

* first pass writing out wpa_supplicant.conf

* manifests directory should be effectivly in the container

* bug fixes Dockerfile and startup.py

need pip for PyYAML.
Specify full path to wpa_passphrase.

* fix bugs in writeWPAConfig

* log entry into writeWPAConfig()

* updated ui resource to include manifests volume

* updated how the wpa_conf secret is written out

* remove extraneous arguments from wpa_supplicant cmd

* fix how wpa config is written to a secret

* pass in correct passphrase to wpa config

* attempt to flush device after it comes up

* wifi deployment to run once network is known

* flush the network adapter after hostapd comes up

* update imagePullPolicy to IfNotPresent for all setup images

* declare image tags that are specific not latest

* remove init() operation from ap.sh

Use configmaps instead.

* use local tag for basic build as with other container images

* updates to interactive yaml configuration

Move turnkey namespace to a separate yaml.
Configure hostapd to pickup configmaps instead of using init()
Add hostapd configmaps to interactive yaml resources

* upgrade hostapd resource to use v0.0.3 tag

* update volume mount of configmaps

* volumeMounts were added to the init container, not hostapd

* cleanup volume mounts

* add soft links to the config files

add configfiles to the conf directory using configmaps in k8s.

* Another pass at configmaps using docker workdir instead of /etc

* use v0.0.4 of hostapd container image

* command arg should be up not reset

* network container improvements

Turn of debug output of shell scripts during the sleep loops
Add a wait command after starting wpa_supplicant

* updated yaml to reflect recent image updates.

* Configure automatic switch from hostapd to wifi

Update ui flow
Enable [up | down | sleep] capability for hostapd
Push wifi config out when hostapd is down/sleeping

* updated to use recent updates to images

* remove spaces in variable assignment statements

* upgraded to use v0.0.6 hostapd image

* we cannot get email from a form that doesn't have email

* use updated ui container

* lighter logging from the sleep loop in ap

* volume needs to be mounted to see the file

* update container version

* container version was the same only k8s config changed

* new app route needs methods - now POST

* update ui container version

* added flask server shutdown when complete

* use nftables instead of (legacy) iptables

* updated turnkey image tag

* Change how ui container runs/exits

Added entrypoint script to manage python process.
Removed run and build scripts in favor of Makefile management of
the same operations.
Updated Dockerfile to reflect changes.

* Include entrypoint in dockerfile.

add debug output to entrypoint script.
access status from the correct location in the container.

* updated ui image tag

* entrypoint is now in correct location and executable.

conditional logic in entrypoint.sh has been updated to correctly
compare values.

* updated ui image tag

* Host pre-configuration setup scripts and content

* updated help

* updated ip addrs to be link-local

* ssids and status should be created if they don't exist

* ImagePullPolich on wifi.yaml should be IfNotPresent

* ui revision update

* documentation

* setup and resolv.conf updated with non link-local address

* Update setup.sh

allow brancher to be set as an env variable before running the script.
It's now possible to set TURNKEY_BRANCH env variable so that the 
script can be piped to sh without ever downloading it

* operations to reset the system to pre configured k3s turnkey

* modified setup process

* explicitly place turnkey files

* replace shell script builds with makefile

* moved ap.sh to the 'standard' entrypoint.sh

* exit with failure code if hostapd is unable to setup interface

* use v0.0.9 of hostapd

+++ network/ap/entrypoint.sh
@@ -4,51 +4,12 @@ set -x
 while getopts h flag
 do
     case "${flag}" in
-        h) echo "ap.sh [up|down|reset|init]"; 
+        h) echo "ap.sh [up|down|reset]"; 
 			exit 0
 			;;
     esac
 done
 
-function init()
-{
-	# configure dhcpcd
-	cat <<-'EOF' >> /etc/dhcpcd.conf
-
-	interface wlan0
-	    static ip_address=192.168.4.1/24
-	    nohook wpa_supplicant 
-	EOF
-
-	# Create a new dnsmasq configuration for this device
-	cat <<-EOF >> /etc/dnsmasq.conf
-	interface=wlan0
-	dhcp-range=192.168.4.2,192.168.4.20,255.255.255.0,24h
-	EOF
-
-	# setup hostapd configuration
-	cat <<-EOF > /etc/hostapd/hostapd.conf
-	interface=wlan0
-	driver=nl80211
-	ssid=ConfigureK3s
-	hw_mode=a
-	channel=44
-	ieee80211d=1
-	country_code=US
-	ieee80211n=1
-	ieee80211ac=1
-	wmm_enabled=0
-	macaddr_acl=0
-	auth_algs=1
-	ignore_broadcast_ssid=0
-	wpa=2
-	wpa_passphrase=rancher-k3s
-	wpa_key_mgmt=WPA-PSK
-	wpa_pairwise=TKIP
-	rsn_pairwise=CCMP
-	EOF
-
-}
 
 function ap_up()
 {
@@ -68,14 +29,20 @@ function ap_up()
         /etc/init.d/dnsmasq systemd-start-resolvconf
     else
         echo dnsmasq cannot be started
-        exit 0
+        exit 1
     fi
 
     # fire up hostapd
     DAEMON_CONF=/etc/hostapd/hostapd.conf
-    /usr/sbin/hostapd -B -P /run/hostapd.pid ${DAEMON_CONF}
+    /usr/sbin/hostapd -B -f /root/hostapd.out ${DAEMON_CONF}
 
     wait $!
+
+	if grep -e "Unable to setup interface" /root/hostapd.out; then
+		exit 1
+	fi 
+	# flush the network adapter after hostapd comes up
+	ip addr flush dev wlan0
 
     # enable ip forwarding 
     echo "1" > /proc/sys/net/ipv4/ip_forward
@@ -93,7 +60,6 @@ function ap_down()
 
 function reset()
 {
-    # TODO: RELOAD CONFIGURATION FILES
     ap_down
 	wait $!
     ap_up
@@ -102,9 +68,8 @@ function reset()
 trap ap_down SIGTERM
 trap reset SIGUSR1
 
-if [ "$1" == "reset" ] || [ "$1" == "init" ] 
+if [ "$1" == "reset" ] 
 then
-	init
 	reset
 elif [ "$1" == "down" ]
 then
@@ -114,7 +79,35 @@ then
 	ap_up
 fi
 
-while true
+# Status
+# up - hostapd service should be up
+# down - exit 
+# sleep - hostapd service should be down, maintain listening loop
+status='up'
+cur=$status
+prev=$status
+until [ "$status" == "down" ]
 do
-    sleep 10
+	status=$(cat /var/lib/rancher/turnkey/status)
+	cur=$status
+	if [ "$status" == "sleep" ] && [ "$cur" != "$prev" ]
+	then
+		# continue logging
+		set -x
+		ap_down
+	elif [ "$status" == "up" ] && [ "$cur" != "$prev" ]
+	then 
+		# continue logging
+		set -x
+		ap_up
+	fi
+	prev=$cur
+
+	sleep 5
+	# stop debugging at the end of the loop
+	set +x
 done
+
+set -x
+# be sure the ap is down when exiting
+ap_down
